requirements.txt
flask>=2.0.0
yfinance>=0.2.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
